+++ src/HeimrichHannotFlatpickrBundle.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (c) 2018 Heimrich & Hannot GmbH
+ *
+ * @author  Thomas Körner <t.koerner@heimrich-hannot.de>
+ * @license http://www.gnu.org/licences/lgpl-3.0.html LGPL
+ */
+
+
+namespace HeimrichHannot\FlatpickrBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class HeimrichHannotFlatpickrBundle extends Bundle
+{
+
+}