+++ src/Resources/assets/js/contao-flatpickr-bundle-init.js
@@ -1,0 +1,3 @@
+import { FlatpickrBundle } from './contao-flatpickr-bundle';
+
+document.addEventListener('DOMContentLoaded', FlatpickrBundle.init);