+++ src/Resources/assets/js/contao-flatpickr-bundle-theme.js
@@ -1,1 +1,2 @@
+import '../../public/assets/flatpickr.css';
 import '../scss/flatpickr-bundle.scss';